Add unit tests for extract helpers
