+++ src/index.ts
@@ -15,6 +15,7 @@ import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js"
 import { EmptyResultSchema } from "@modelcontextprotocol/sdk/types.js";
 // 自作モジュールからの関数インポート
 import { getAllTools, callTool } from './mcp/client.js';
+import { extractToolCalls, extractCodeBlocks } from './utils/extract.js';
 
 // MCPサーバーマネージャー
 let mcpServerManager: MCPServerManager | null = null;
@@ -625,73 +626,6 @@ export async function setupWizard(): Promise<void> {
   }
 }
 
-// コンテンツからツール呼び出しを抽出する関数
-function extractToolCalls(content: string): { tool: string; args: any }[] {
-  const calls: { tool: string; args: any }[] = [];
-  
-  // 正規表現でツール呼び出しパターンを検索
-  // パターン1: tools/call filesystem ls {"path":"/"}
-  const regex1 = /tools\/call\s+(\w+)\s+(\w+)\s+({.*?})/g;
-  let match;
-  
-  while ((match = regex1.exec(content)) !== null) {
-    try {
-      const server = match[1]; // サーバー名（例: filesystem）
-      const tool = match[2];   // ツール名（例: ls）
-      const argsStr = match[3]; // 引数（例: {"path":"/"})
-      
-      // 完全なツール名を構築
-      const fullToolName = `${server}.${tool}`;
-      
-      // JSON文字列をパース
-      const args = JSON.parse(argsStr);
-      
-      calls.push({ tool: fullToolName, args });
-    } catch (e) {
-      if (globalLogLevel !== 'quiet') {
-        console.warn('ツール呼び出しの解析エラー:', e instanceof Error ? e.message : String(e));
-      }
-    }
-  }
-  
-  // パターン2: tools/call ls {"path":"/"}（サーバー名なし）
-  const regex2 = /tools\/call\s+(\w+)\s+({.*?})/g;
-  
-  while ((match = regex2.exec(content)) !== null) {
-    // regex1ですでに処理したマッチはスキップ
-    const fullMatch = match[0];
-    if (regex1.test(fullMatch)) continue;
-    
-    try {
-      const tool = match[1];   // ツール名（例: ls）
-      const argsStr = match[2]; // 引数（例: {"path":"/"})
-      
-      // JSON文字列をパース
-      const args = JSON.parse(argsStr);
-      
-      calls.push({ tool, args });
-    } catch (e) {
-      if (globalLogLevel !== 'quiet') {
-        console.warn('ツール呼び出しの解析エラー:', e instanceof Error ? e.message : String(e));
-      }
-    }
-  }
-  
-  return calls;
-}
-
-// マークダウンからコードブロックを抽出
-function extractCodeBlocks(text: string): string[] {
-  const codeBlockRegex = /```(?:javascript|js)?\n([\s\S]*?)```/g;
-  const blocks: string[] = [];
-  let match: RegExpExecArray | null;
-  
-  while ((match = codeBlockRegex.exec(text)) !== null) {
-    blocks.push(match[1]);
-  }
-  
-  return blocks;
-}
 
 // 指定されたツールをサポートするサーバーIDを探す
 async function findServerForTool(
@@ -720,3 +654,4 @@ async function findServerForTool(
   }
   return undefined;
 }
+
